Adjust description declaration to the inside of the AnalyticsHubTimeRange
